lint

+++ src/anki.rs
@@ -159,13 +159,13 @@ impl AppState {
             .for_each(|(i, anki_s)| {
                 let req: Request<UpdateNoteParams> = match &anki_s.filename.clone() {
                     Some(filename) => into_update_note_req(
-                        note_ids[i] as u128,
+                        note_ids[i],
                         &config.fields,
                         anki_s,
                         filename.to_string(),
                     ),
                     None => {
-                        into_update_only_sentence_req(note_ids[i] as u128, &config.fields, &anki_s)
+                        into_update_only_sentence_req(note_ids[i], &config.fields, &anki_s)
                     }
                 };
                 requests_vec.push(req);
